refactor parser packages
