+++ internal/adapter/parser/aviation/aviation_parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package aviation
 
 import (
 	"caatsm/internal/domain"
