+++ internal/adapter/parser/schedule/schedule_parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package schedule
 
 import (
 	"strings"
